+++ CHANGELOG.md
@@ -4,10 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.21.1-4.0.11] - Unreleased
+### Added
+- Configuration screen, accessible from the Mods menu
+- Config option to make crouching/sneaking disable item magnet effects (enabled by default)
 ### Fixed
 - Possibly fixed `ConcurrentModificationException` crash with magnet pull trait effects [#762]
 - Keybindings for tooltips not working in some situations [#761]
 - Most blocks should now interact correctly with pipes and other item transfer methods from other mods [#735]
+- Miscellaneous bugs with item magnet effects
 
 ## [1.21.1-4.0.10] - 2024-12-20
 ### Fixed